+++ packages/autotool/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @llamaindex/autotool
 
+## 7.0.5
+
+### Patch Changes
+
+- llamaindex@0.10.5
+
 ## 7.0.4
 
 ### Patch Changes