+++ packages/providers/vercel/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @llamaindex/vercel
 
+## 0.1.4
+
+### Patch Changes
+
+- Updated dependencies [9b2e25a]
+  - @llamaindex/core@0.6.4
+
 ## 0.1.3
 
 ### Patch Changes