+++ packages/workflow/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @llamaindex/workflow
 
+## 1.1.1
+
+### Patch Changes
+
+- Updated dependencies [9b2e25a]
+  - @llamaindex/core@0.6.4
+  - @llamaindex/env@0.1.30
+
 ## 1.1.0
 
 ### Minor Changes